moved tc-ui server side stuff all into tc-ui package
