+++ packages/tc-ui/handlers/index.js
@@ -1,0 +1,5 @@
+const { getViewHandler } = require('./view');
+const { getEditHandler } = require('./edit');
+const { getDeleteHandler } = require('./delete');
+
+module.exports = { getViewHandler, getEditHandler, getDeleteHandler };
