+++ packages/tc-ui/server.js
@@ -5,7 +5,72 @@ const { ApiClient } = require('./lib/api-client');
 const { getSchemaSubset } = require('./lib/get-schema-subset');
 const { getDataTransformers } = require('./lib/get-data-transformers');
 const { getPageRenderer, FormError } = require('./page');
+const {
+	getViewHandler,
+	getEditHandler,
+	getDeleteHandler,
+} = require('./handlers');
 const messages = require('./components/messages');
+
+const getCMS = ({
+	logger,
+	apiBaseUrl,
+	apiHeaders,
+	Header,
+	Footer,
+	customComponents,
+	customTypeMappings,
+}) => {
+	const assignComponent = componentAssigner({
+		customComponents,
+		customTypeMappings,
+	});
+
+	const graphqlBuilder = type => graphqlQueryBuilder(type, assignComponent);
+
+	const { handleError, renderPage } = getPageRenderer({
+		Header,
+		Footer,
+	});
+
+	const getApiClient = event =>
+		new ApiClient({
+			event,
+			graphqlBuilder,
+			logger,
+			apiBaseUrl,
+			apiHeaders,
+		});
+
+	const { formDataToRest, formDataToGraphQL } = getDataTransformers(
+		assignComponent,
+	);
+	const { handler: viewHandler, render: viewRender } = getViewHandler({
+		getApiClient,
+		getSchemaSubset,
+		handleError,
+		assignComponent,
+		renderPage,
+	});
+	return {
+		view: viewHandler,
+		edit: getEditHandler({
+			getApiClient,
+			getSchemaSubset,
+			handleError,
+			renderPage,
+			assignComponent,
+			formDataToRest,
+			formDataToGraphQL,
+		}).handler,
+		delete: getDeleteHandler({
+			getApiClient,
+			handleError,
+			viewRender,
+			logger,
+		}).handler,
+	};
+};
 
 module.exports = {
 	primitives,
@@ -17,6 +82,7 @@ module.exports = {
 	FormError,
 	getDataTransformers,
 	components: { ...messages },
+	getCMS,
 };
 
 /*
